+++ src/components/Button/CHANGELOG.md
@@ -8,3 +8,4 @@
   center` to the button styles.
 - 2025-09-29: Make the prop `size` support `xs` and `xl` values & use tokens
   from ThemeProvider.
+- 2025-10-04: Add `ghost` variant. Code formatting improvement.